[WEEK1 - STEP3] refactor : main과 step2 파일 분리
