+++ CodeStarterCamp_Week1/step2.swift
@@ -1,0 +1,22 @@
+//
+//  step2.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by soyounglee on 2022/02/11.
+//
+
+import Foundation
+
+// 1. 아이스크림 부분을 그리는 함수
+func drawIceCreamTop() {
+  for _ in 0..<8 {
+    print("***********\n")
+  }
+}
+
+// 2. 막대기(손잡이) 부분을 그리는 함수
+func drawIceCreamBar() {
+  for _ in 0..<4 {
+    print("    | |\n")
+  }
+}
